requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiomultiprocess==0.9.0
aiosignal==1.3.1
aiosqlite==0.17.0
aniso8601==9.0.1
anyio==3.6.2
apparmor==3.0.8
appdirs==1.4.4
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.10.3
backcall==0.2.0
backdoor-factory==0.0.0
backoff==2.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.2
bidict==0.22.1
binwalk==2.3.3
blinker==1.5
Bottleneck==1.3.5
Brotli==1.0.9
cairocffi==1.4.0
capstone==4.0.2
censys==2.1.9
Cerberus==1.3.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dhcpig==0.0.0
dicttoxml==1.7.15
distro==1.8.0
Django==3.2.19
django-filter==23.1
django-guardian==2.4.0
djangorestframework==3.14.0
dnslib==0.9.23
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dsinternals==1.2.4
EasyProcess==1.1
et-xmlfile==1.0.1
eventlet==0.33.1
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.92.0
flasgger==0.9.5
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.2
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
future==0.18.2
fuzzywuzzy==0.18.0
GeoIP==1.3.2
geoip2==2.9.0
gevent==22.10.2
gevent-websocket==0.10.1
gpg==1.18.0
greenlet==2.0.2
gyp==0.1
h11==0.14.0
h2==4.1.0
hashID==3.1.4
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
hyperframe==6.0.0
idna==3.3
imageio==2.4.1
impacket==0.11.0
importlib-metadata==4.12.0
iniconfig==1.1.1
instaloader==4.11
invoke==2.0.0
ipdb==0.13.9
ipython==8.5.0
itsdangerous==2.1.2
itypes==1.2.0
jdcal==1.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.10.3
kaitaistruct==0.10
kiwisolver==0.0.0
ldap3==2.9.1
ldapdomaindump==0.9.3
LibAppArmor==3.0.8
limits==2.8.0
logutils==0.3.5
lxml==4.9.2
lz4==4.0.2+dfsg
Mako==1.2.4.dev0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mate-menu==22.4.1
mate-tweak==22.10.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
maxminddb==2.1.0
mdurl==0.1.2
mechanize==0.4.8
mistune0==0.8.4
mitmproxy==8.1.1
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
munkres==1.1.4
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
notify2==0.3
ntpsec==1.2.2
numexpr==2.8.4
numpy==1.24.2
nyx==2.1.0
odfpy==1.4.2
olefile==0.46
onboard==1.4.1
onionshare==2.6
onionshare-cli==2.6
openpyxl==3.0.9
outcome==1.2.0
packaging==25.0
pandas==1.5.3
paramiko==2.12.0
parso==0.8.3
passlib==1.7.4
pcapy==0.11.5.dev0
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pluggy==1.0.0+repack
pluginbase==1.0.1
ply==3.11
prettytable==3.6.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.0.0
py==1.11.0
pyasn1==0.4.8
pycairo==1.20.1
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.4
pyee==9.0.4
pygeoip==0.3.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
pymongo==3.11.0
pymssql==2.2.2
PyNaCl==1.5.0
pynvim==0.4.2
pyodbc==4.0.34
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypng==0.20220715.0
pyppeteer==1.0.1
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyrtlsdr==0.2.91
pyserial==3.5
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==0.21.0
python-engineio==4.3.4
python-Levenshtein==0.12.2
python-libnmap==0.7.2
python-magic==0.4.26
python-magic-ahupp==0.4.13
python-memcached==1.58
python-multipart==0.0.5
python-socketio==5.7.2
python-xlib==0.33
pytz==2022.7.1
pyusb==1.2.1.post2
PyVirtualDisplay==2.2
pyxdg==0.28
PyYAML==6.0
qrcode==7.4.2
redis==4.3.4
requests==2.28.1
requests-futures==1.0.0
requests-toolbelt==0.10.1
retrying==1.3.3
rfc3986==1.5.0
rfc3987==1.3.8
rfcat==2.0.1
rich==13.3.1
rocket==1.3.0
rq==1.13.0
rtlsdr-scanner==1.3.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scapy==2.5.0
selenium==4.15.2
setproctitle==1.3.1
sherlock-project==0
shodan==1.28.0
simplejson==3.18.3
six==1.16.0
slowapi==0.1.4
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
specan==0.0.0
SQLAlchemy==1.4.46
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.26.1
stem==1.8.1
sympy==1.11.1
tables==3.7.0
tabulate==0.8.9
termcolor==1.1.0
texttable==1.6.7
theHarvester==4.4.3
toml==0.10.2
torbrowser-launcher==0.3.6
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.13.2
ufoLib2==0.14.0
ufw==0.36.2
ujson==5.7.0
unicodedata2==15.0.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
urwid==2.1.2
uvicorn==0.17.6
uvloop==0.17.0
validators==0.20.0
vinetto==0.8.0
visvis==1.13.0
wafw00f==2.2.0
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.4
websploit==4.0.4
Werkzeug==2.2.2
wfuzz==3.1.0
wifite==2.6.6
wrapt==1.14.1
wsproto==1.2.0
wxPython==4.2.0
xcffib==0.11.1
xdg==5
XlsxWriter==3.0.2
yara-python==4.5.1
yarl==1.8.2
yt-dlp==2025.4.30
zipp==1.0.0
zope.event==4.4
zope.interface==5.5.2
zstandard==0.20.0
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-dotenv
Pillow
celery[redis]
django-cors-headers
django-environ
graphviz
